power_bi_sales_reporting_automation/requirements.txt
prefect==3.6.19
requests
msal
pyodbc
PyMuPDF
python-dotenv